requirements.txt
Django>=5.0
gunicorn
psycopg2-binary
paho-mqtt
dj-database-url
python-dotenv
